requirements.txt
psycopg2==2.8.5
gunicorn==20.0.4
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
PyJWT==1.7.1
django-filter==2.3.0
django-rest-swagger==2.2.0
django-rest-auth==0.9.5
django-cors-headers==3.4.0
django-select2==7.4.2
django-admin-sortable2==0.7.6
django-cacheback==2.0.0
django-picklefield==3.0.1
django-defender==0.9.1
django-json-widget==1.1.1
jsonfield==3.1.0
sentry-sdk==1.4.3
mock==4.0.2
PyYAML==5.3.1
factory-boy==2.12.0
django-redis-cache==2.1.3
redis==3.5.3
rq==1.9.0
six==1.14.0
django-rq==2.4.1
pytz==2020.1
urllib3==1.25.9
yamjam==0.1.7
pytest==5.4.2
pytest-django==3.9.0
paho-mqtt==1.5.1
python-geohash==0.8.5
Django==2.2.12
